+++ src/main/java/com/remoteLaboratory/service/Impl/CourseServiceImpl.java
@@ -69,6 +69,9 @@ public class CourseServiceImpl implements CourseService {
                 if(!loginUser.getUserType().equals(Constants.USER_TYPE_ADMIN) && !course.getTeacherId().equals(loginUser.getId())) {
                     throw new BusinessException(Messages.CODE_50200);
                 }
+                if(course.getStatus().equals(1)) {
+                    throw new BusinessException(Messages.CODE_40010, "课程进行中不能执行删除操作");
+                }
                 courseRepository.delete(id);
                 LogRecord logRecord = new LogRecord();
                 logRecord.setType("删除");
